Consistently use async
